speed up fetch tests by giving them their own repo-fixture (#450)
