+++ git-repository/tests/fixtures/generated-archives/.gitignore
@@ -1,3 +1,4 @@
 /make_worktree_repo.tar.xz
 /make_remote_repos.tar.xz
+/make_fetch_repos.tar.xz
 /make_core_worktree_repo.tar.xz
