+++ git-repository/tests/fixtures/make_remote_repos.sh
@@ -97,19 +97,6 @@ git clone --shared base clone
   git remote add myself .
 )
 
-git clone --shared base clone-as-base-with-changes
-(cd clone-as-base-with-changes
-  touch new-file
-  git add new-file
-  git commit -m "add new-file"
-  git tag -m "new-file introduction" v1.0
-)
-
-git clone --shared base two-origins
-(cd two-origins
-  git remote add changes-on-top-of-origin $PWD/../clone-as-base-with-changes
-)
-
 git clone --shared base push-default
 (cd push-default
